+++ src/Utils/DatabaseUtils.php
@@ -8,6 +8,16 @@ use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\DataAbstractionLayer\DefinitionInstanceRegistry;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 
+/**
+ * This class provides utility methods to directly interact with the database. It has build in
+ * caching to prevent multiple database queries for the same data within one command
+ * execution / request.
+ *
+ * This class is designed to be used through the FixtureHelper, using:
+ * ```php
+ * $this->helper->Database()->……();
+ * ```
+ */
 readonly class DatabaseUtils
 {
     public function __construct(