+++ src/Utils/PaymentMethodUtils.php
@@ -12,6 +12,15 @@ use Shopware\Core\Framework\DataAbstractionLayer\EntityRepository;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
 
+/**
+ * This class provides utility methods to work with payment methods. It has build in caching to prevent
+ * multiple database queries for the same data within one command execution / request.
+ *
+ * This class is designed to be used through the FixtureHelper, using:
+ * ```php
+ * $this->helper->PaymentMethod()->……();
+ * ```
+ */
 readonly class PaymentMethodUtils
 {
     /**
@@ -22,16 +31,23 @@ readonly class PaymentMethodUtils
     ) {
     }
 
+    /**
+     * Return the default invoice payment method of shopware.
+     */
     public function getInvoicePaymentMethod(): ?PaymentMethodEntity
     {
-        $criteria = (new Criteria())->addFilter(
-            new EqualsFilter('handlerIdentifier', InvoicePayment::class),
-        )->setLimit(1);
+        return once(function (): ?PaymentMethodEntity {
+            $criteria = (new Criteria())->addFilter(
+                new EqualsFilter('handlerIdentifier', InvoicePayment::class),
+            )->setLimit(1);
 
-        $paymentMethod = $this->paymentMethodRepository
-            ->search($criteria, Context::createDefaultContext())
-            ->first();
+            $criteria->setTitle(sprintf('%s::%s()', __CLASS__, __FUNCTION__));
 
-        return $paymentMethod instanceof PaymentMethodEntity ? $paymentMethod : null;
+            $paymentMethod = $this->paymentMethodRepository
+                ->search($criteria, Context::createDefaultContext())
+                ->first();
+
+            return $paymentMethod instanceof PaymentMethodEntity ? $paymentMethod : null;
+        });
     }
 }